Replace gulp with npm scripts
gulp-cli hasn't been updated since 2019 and is causing a lot of
alerts about insecure JavaScript dependencies; what if we could get rid
of all those dependencies instead of working out how to update them?

This commit replaces our gulp workflow with a couple of npm scripts.

Notes
=====

Our gulp workflow was pretty simple to begin with, but this version is
missing a few things:

1. no difference between building for production or development
2. no watch script
3. no pretty output with emojis

However I think sourcemaps have pretty much removed the need for 1., and
we can add 2. at a later date if there is demand for it.

I miss 3. :sniff:, but I think it is a price worth paying for so many
fewer dependencies.

Why not use invoke?
===================

Over in Python land we're currently migrating our makefiles to `invoke`;
I considered using invoke instead of npm scripts but then we would have
to install digitalmarketplace-developer-tools in our Docker images, and
that meant more complexity that I couldn't be bothered with. But we
could do it, with the same commands, without too much hassle.